requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.8.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-channels==0.7.0
django-debug-toolbar==2.2
django-jsonfield==1.4.0
django-modelcluster==5.0.1
django-multiupload2==0.5.3
django-paypal==1.0.0
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
l18n==2018.5
msgpack==0.6.2
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkg-resources==0.0.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.48.0
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9
wcwidth==0.1.9
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.8
zope.interface==5.1.0
